Use a lua expression as the tabline variable (and other enhancements)
Use a lua expression as the tabline variable. This means we don't have
to rely on autocmd events to do a rebuild of the tabline since we can
just rely on nvim's built-in re-eval logic. The tabline will only be
redrawn when neovim knows something has changed, which includes saving
buffers (so they are no longer modified) on non-visible buffers (fixes

Because we aren't repeatedly updating the tabline option text, the
start/help screen when vim is first opened is no longer immediately
erased which fixes #22.

Loading buftabline will now occur on VimEnter if it is autoloaded at
startup, but will occur synchronously if re-loaded or if it's
lazy-initialized.

Auto-hide is now checked on a separate autocmd and is no longer part of
the build() function.

Tests updated to no longer need to wait for the scheduler since we
aren't scheduling the command to update the tabline anymore.